Updated ReviewCard.jsx, Implemented half star rating system

+++ app/models/rating.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey, DateTime, func
+from sqlalchemy import Column, Integer, Float, String, ForeignKey, DateTime, func
 from sqlalchemy.orm import relationship
 from app.db.database import Base
 
@@ -8,7 +8,7 @@ class Rating(Base):
     id = Column(Integer, primary_key=True, index=True, autoincrement=True)
     user_id = Column(Integer, ForeignKey("users.id"))
     album_id = Column(Integer, ForeignKey("albums.id"))
-    score = Column(Integer, nullable=False)
+    score = Column(Float, nullable=False)
     review = Column(String)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
